requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.8
djangorestframework==3.12.4
geographiclib==1.52
geopy==2.2.0
iniconfig==1.1.1
mock==4.0.3
numpy==1.21.3
packaging==21.2
pandas==1.3.4
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
